some minor changes
